feat: fix calender ui
